[EXSHAPP-0436]: Refactor Currency Calculations – BigDecimal Migration and Remainder Distribution (#442)

+++ domain/src/main/kotlin/es/pedrazamiguez/expenseshareapp/domain/repository/UserRepository.kt
@@ -1,5 +1,7 @@
 package es.pedrazamiguez.expenseshareapp.domain.repository
 
+import java.math.BigDecimal
+
 interface UserRepository {
-    suspend fun getUserBalance(userId: String): Double
+    suspend fun getUserBalance(userId: String): BigDecimal
 }
